+++ examples/kyb-app/src/domains/collection-flow/collection-flow.api.ts
@@ -1,0 +1,67 @@
+import { request } from '@app/common/utils/request';
+import {
+  AuthorizeDto,
+  FlowData,
+  GetActiveWorkflowDto,
+  GetSessionDto,
+  TFlowConfiguration,
+  TFlowStep,
+  TUser,
+  UpdateFlowDto,
+} from '@app/domains/collection-flow/types';
+
+export const authorizeUser = async (dto: AuthorizeDto): Promise<TUser> => {
+  const result = await request
+    .post('collection-flow/authorize', {
+      json: {
+        email: dto.email,
+        flowType: import.meta.env.VITE_KYB_DEFINITION_ID,
+      },
+    })
+    .json<TUser>();
+
+  return {
+    id: result.id,
+    email: result.email,
+    businessId: result.businesses.at(-1).id,
+  } as TUser;
+};
+
+export const getFlowSession = async (query: GetSessionDto) => {
+  // Uisng same endpoint as used for auth till session mechanism will be implemented in backend for collection flow
+  return authorizeUser(query);
+};
+
+export const fetchCollectionFlowSchema = async (): Promise<TFlowStep[]> => {
+  const result = await request
+    .get(`collection-flow/configuration`, {
+      searchParams: {
+        flowType: import.meta.env.VITE_KYB_DEFINITION_ID as string,
+      },
+    })
+    .json<TFlowConfiguration>();
+
+  return result.steps;
+};
+
+export const fetchActiveWorkflow = async (dto: GetActiveWorkflowDto): Promise<FlowData> => {
+  const result = await request
+    .get('collection-flow/active-flow', {
+      searchParams: {
+        endUserId: dto.endUserId,
+        flowType: import.meta.env.VITE_KYB_DEFINITION_ID as string,
+      },
+    })
+    .json<{ result: FlowData }>();
+
+  return result.result;
+};
+
+export const updateFlow = async (dto: UpdateFlowDto) => {
+  const { flowId, ...payload } = dto;
+  await request.put(`collection-flow/${flowId}`, { json: payload });
+};
+
+export const startFlow = async (flowId: string) => {
+  await request.post(`collection-flow/finish/${flowId}`);
+};