+++ examples/kyb-app/src/pages/CollectionFlow/components/organisms/KYBView/views/HeadquartersView/HeadquartersView.tsx
@@ -8,17 +8,19 @@ import { useHeadquartersSchema } from '@app/pages/CollectionFlow/components/orga
 import { HeadquartersContext } from '@app/pages/CollectionFlow/components/organisms/KYBView/views/HeadquartersView/types';
 import { WorkflowFlowData } from '@app/domains/workflows/flow-data.type';
 import { DynamicForm } from '@ballerine/ui';
+import { useViewSchemas } from '@app/pages/CollectionFlow/components/organisms/KYBView/hooks/useViewSchemas';
 
 export const HeadquartersView = () => {
   const { context, update, saveAndPerformTransition } = useViewState<WorkflowFlowData>();
-  const { schema } = useHeadquartersSchema(context.flowData.headquarters, headquartersSchema);
+  const { formSchema, uiSchema } = useViewSchemas();
+  const { schema } = useHeadquartersSchema(context.flowData.headquarters, formSchema);
 
   return (
     <AppShell.FormContainer header={<ViewHeader />}>
       <DynamicForm<HeadquartersContext>
         className="max-w-[384px]"
         schema={schema}
-        uiSchema={headquartersUISchema}
+        uiSchema={uiSchema}
         formData={context.flowData.headquarters}
         onChange={values => void update(values)}
         onSubmit={values => void saveAndPerformTransition(values)}